Unified file creation procedure
